[ADD][REF] trpc, core: clena up and added trpc

+++ src/lib/data-table/utils.ts
@@ -18,7 +18,7 @@ export function formatDate(
 	opts: Intl.DateTimeFormatOptions = {}
 ) {
 	return new Intl.DateTimeFormat("en-US", {
-		month: opts.month ?? "long",
+		month: opts.month ?? "short",
 		day: opts.day ?? "numeric",
 		year: opts.year ?? "numeric",
 		...opts,
